separate module fmt for debug output
